refactor + encapsultae ImmuTableRow creation for ViewsVisitors and updated tests
