+++ WordPress/WordPressTest/ViewRelated/Stats/Insights/SiteStatsInsightViewModelTests.swift
@@ -7,7 +7,7 @@ class SiteStatsInsightsViewModelTests: XCTestCase {
     /// The standard api result for a normal user
     func testSummarySplitIntervalData14DaysBasecase() throws {
         // Given statsSummaryTimeIntervalData with 14 days data
-        guard let statsSummaryTimeIntervalData = try! createStatsSummaryTimeIntervalData(fileName: "stats-visits-day-14.json") else {
+        guard let statsSummaryTimeIntervalData = try! StatsMockDataLoader.createStatsSummaryTimeIntervalData(fileName: "stats-visits-day-14.json") else {
             XCTFail("Failed to create statsSummaryTimeIntervalData")
             return
         }
@@ -19,7 +19,7 @@ class SiteStatsInsightsViewModelTests: XCTestCase {
     /// The api result for a new user that has full data for this week but a partial dataset for the previous week
     func testSummarySplitIntervalData11Days() throws {
         // Given statsSummaryTimeIntervalData with 11 days data
-        guard let statsSummaryTimeIntervalData = try! createStatsSummaryTimeIntervalData(fileName: "stats-visits-day-11.json") else {
+        guard let statsSummaryTimeIntervalData = try! StatsMockDataLoader.createStatsSummaryTimeIntervalData(fileName: "stats-visits-day-11.json") else {
             XCTFail(Constants.failCreateStatsSummaryTimeIntervalData)
             return
         }
@@ -31,7 +31,7 @@ class SiteStatsInsightsViewModelTests: XCTestCase {
     /// The api result for a new user that has an incomplete dataset for this week and no data for prev week
     func testSummarySplitIntervalData4Days() throws {
         // Given statsSummaryTimeIntervalData with 4 days data
-        guard let statsSummaryTimeIntervalData = try! createStatsSummaryTimeIntervalData(fileName: "stats-visits-day-4.json") else {
+        guard let statsSummaryTimeIntervalData = try! StatsMockDataLoader.createStatsSummaryTimeIntervalData(fileName: "stats-visits-day-4.json") else {
             XCTFail(Constants.failCreateStatsSummaryTimeIntervalData)
             return
         }
@@ -56,51 +56,6 @@ class SiteStatsInsightsViewModelTests: XCTestCase {
                 XCTAssertEqual(prevWeek.summaryData.first?.periodStartDate, Calendar.autoupdatingCurrent.date(from: DateComponents(year: 2019, month: 2, day: 8)))
             }
         }
-    }
-
-    func convertStringToDictionary(text: String) -> [String: AnyObject]? {
-        if let data = text.data(using: .utf8) {
-            do {
-                let json = try JSONSerialization.jsonObject(with: data, options: .mutableContainers) as? [String: AnyObject]
-                return json
-            } catch {
-                print("Something went wrong")
-            }
-        }
-        return nil
-    }
-
-    func createStatsSummaryTimeIntervalData(fileName: String) throws -> StatsSummaryTimeIntervalData? {
-        let feb21 = DateComponents(year: 2019, month: 2, day: 21)
-        let date = Calendar.autoupdatingCurrent.date(from: feb21)!
-
-        let data = try load(fileName: fileName, fromBundle: Bundle(for: type(of: self)))
-        let str = String(decoding: data, as: UTF8.self)
-        let jsonDictionary = convertStringToDictionary(text: str)
-
-        guard let json = jsonDictionary else {
-            return nil
-        }
-
-        guard let statsSummaryTimeIntervalData = StatsSummaryTimeIntervalData(date: date,
-                period: .day,
-                jsonDictionary: json) else {
-            XCTFail()
-            return nil
-        }
-
-        return statsSummaryTimeIntervalData
-    }
-
-    func load(fileName: String, fromBundle bundle: Bundle) throws -> Data {
-        let fileUrl = URL(fileURLWithPath: fileName)
-        let baseName = fileUrl.deletingPathExtension().path
-        let ext = fileUrl.pathExtension
-
-        guard let path = bundle.path(forResource: baseName, ofType: ext) else { XCTFail("Could not find file: \(fileName)"); throw NSError() }
-        let url = URL(fileURLWithPath: path)
-        guard let data = try? Data(contentsOf: url, options: [.mappedIfSafe]) else { XCTFail("Could not load data from file: \(fileName)"); throw NSError() }
-        return data
     }
 }
